WIP: magento events, debug

+++ extension/src/computed_functions_filter.c
@@ -1,3 +1,5 @@
+#include "quanta_mon.h"
+
 void hp_fill_monitored_functions(char **function_names) {
  if (function_names[0] != NULL) return;
  function_names[0] = "Magento\\Framework\\App\\Bootstrap::run";
@@ -20,12 +22,13 @@ void hp_fill_monitored_functions(char **function_names) {
  function_names[17] = "Magento\\Backend\\Controller\\Adminhtml\\Cache\\FlushAll::execute";
  function_names[18] = "Magento\\Framework\\App\\Cache\\TypeList::cleanType";
  function_names[19] = "Magento\\Backend\\Controller\\Adminhtml\\Cache\\FlushSystem::execute";
- function_names[20] = "";
+ function_names[20] = "Magento\\Indexer\\Model\\Indexer::reindexAll";
  function_names[21] = "";
  function_names[22] = NULL;
 }
 
-int hp_match_monitored_function(const char* function_name, const char *class_name) {
+int hp_match_monitored_function(const char* function_name, zend_execute_data* data TSRMLS_DC) {
+ ++hp_globals.internal_match_counters.total;
  if (function_name[0] == 'g') {
   if (function_name[1] == 'e') {
    if (function_name[2] == 't') {
@@ -39,7 +42,11 @@ int hp_match_monitored_function(const char* function_name, const char *class_nam
            if (function_name[10] == 'm') {
             if (function_name[11] == 'e') {
              if (function_name[12] == '\0') {
+              ++hp_globals.internal_match_counters.function;
+              const char *class_name = hp_get_class_name(data TSRMLS_CC);
+              if (!class_name) return -1;
               if (!strcmp(class_name, "Magento\\Framework\\App\\Request\\Http")) return 4;
+              ++hp_globals.internal_match_counters.class_unmatched;
               return -1;
              }
              return -1;
@@ -81,7 +88,11 @@ int hp_match_monitored_function(const char* function_name, const char *class_nam
                    if (function_name[18] == 'k') {
                     if (function_name[19] == 's') {
                      if (function_name[20] == '\0') {
+                      ++hp_globals.internal_match_counters.function;
+                      const char *class_name = hp_get_class_name(data TSRMLS_CC);
+                      if (!class_name) return -1;
                       if (!strcmp(class_name, "Magento\\Framework\\View\\Page\\Builder")) return 9;
+                      ++hp_globals.internal_match_counters.class_unmatched;
                       return -1;
                      }
                      return -1;
@@ -137,7 +148,11 @@ int hp_match_monitored_function(const char* function_name, const char *class_nam
            if (function_name[10] == 's') {
             if (function_name[11] == 'e') {
              if (function_name[12] == '\0') {
+              ++hp_globals.internal_match_counters.function;
+              const char *class_name = hp_get_class_name(data TSRMLS_CC);
+              if (!class_name) return -1;
               if (!strcmp(class_name, "Magento\\Framework\\App\\Response\\Http\\Interceptor")) return 12;
+              ++hp_globals.internal_match_counters.class_unmatched;
               return -1;
              }
              return -1;
@@ -177,7 +192,11 @@ int hp_match_monitored_function(const char* function_name, const char *class_nam
            if (function_name[10] == 'c') {
             if (function_name[11] == 'k') {
              if (function_name[12] == '\0') {
+              ++hp_globals.internal_match_counters.function;
+              const char *class_name = hp_get_class_name(data TSRMLS_CC);
+              if (!class_name) return -1;
               if (!strcmp(class_name, "Magento\\Framework\\View\\Layout")) return 15;
+              ++hp_globals.internal_match_counters.class_unmatched;
               return -1;
              }
              return -1;
@@ -215,7 +234,11 @@ int hp_match_monitored_function(const char* function_name, const char *class_nam
          if (function_name[8] == 'z') {
           if (function_name[9] == 'e') {
            if (function_name[10] == '\0') {
+            ++hp_globals.internal_match_counters.function;
+            const char *class_name = hp_get_class_name(data TSRMLS_CC);
+            if (!class_name) return -1;
             if (!strcmp(class_name, "Magento\\Framework\\Interception\\Config\\Config")) return 3;
+            ++hp_globals.internal_match_counters.class_unmatched;
             return -1;
            }
            return -1;
@@ -245,7 +268,11 @@ int hp_match_monitored_function(const char* function_name, const char *class_nam
      if (function_name[4] == 't') {
       if (function_name[5] == 'e') {
        if (function_name[6] == '\0') {
+        ++hp_globals.internal_match_counters.function;
+        const char *class_name = hp_get_class_name(data TSRMLS_CC);
+        if (!class_name) return -1;
         if (!strcmp(class_name, "Magento\\Framework\\App\\Bootstrap")) return 1;
+        ++hp_globals.internal_match_counters.class_unmatched;
         return -1;
        }
        if (function_name[6] == 'A') {
@@ -260,7 +287,11 @@ int hp_match_monitored_function(const char* function_name, const char *class_nam
                 if (function_name[15] == 'o') {
                  if (function_name[16] == 'n') {
                   if (function_name[17] == '\0') {
+                   ++hp_globals.internal_match_counters.function;
+                   const char *class_name = hp_get_class_name(data TSRMLS_CC);
+                   if (!class_name) return -1;
                    if (!strcmp(class_name, "Magento\\Framework\\App\\Bootstrap")) return 2;
+                   ++hp_globals.internal_match_counters.class_unmatched;
                    return -1;
                   }
                   return -1;
@@ -304,7 +335,11 @@ int hp_match_monitored_function(const char* function_name, const char *class_nam
         if (function_name[7] == 'p') {
          if (function_name[8] == 'e') {
           if (function_name[9] == '\0') {
+           ++hp_globals.internal_match_counters.function;
+           const char *class_name = hp_get_class_name(data TSRMLS_CC);
+           if (!class_name) return -1;
            if (!strcmp(class_name, "Magento\\Framework\\App\\Cache\\TypeList")) return 18;
+           ++hp_globals.internal_match_counters.class_unmatched;
            return -1;
           }
           return -1;
@@ -333,9 +368,13 @@ int hp_match_monitored_function(const char* function_name, const char *class_nam
       if (function_name[5] == 't') {
        if (function_name[6] == 'e') {
         if (function_name[7] == '\0') {
+         ++hp_globals.internal_match_counters.function;
+         const char *class_name = hp_get_class_name(data TSRMLS_CC);
+         if (!class_name) return -1;
          if (!strcmp(class_name, "PDOStatement")) return 16;
          if (!strcmp(class_name, "Magento\\Backend\\Controller\\Adminhtml\\Cache\\FlushAll")) return 17;
          if (!strcmp(class_name, "Magento\\Backend\\Controller\\Adminhtml\\Cache\\FlushSystem")) return 19;
+         ++hp_globals.internal_match_counters.class_unmatched;
          return -1;
         }
         return -1;
@@ -361,8 +400,12 @@ int hp_match_monitored_function(const char* function_name, const char *class_nam
        if (function_name[6] == 'c') {
         if (function_name[7] == 'h') {
          if (function_name[8] == '\0') {
+          ++hp_globals.internal_match_counters.function;
+          const char *class_name = hp_get_class_name(data TSRMLS_CC);
+          if (!class_name) return -1;
           if (!strcmp(class_name, "Magento\\Framework\\App\\FrontController\\Interceptor")) return 7;
           if (!strcmp(class_name, "Magento\\Framework\\App\\Action\\Action")) return 8;
+          ++hp_globals.internal_match_counters.class_unmatched;
           return -1;
          }
          return -1;
@@ -394,11 +437,47 @@ int hp_match_monitored_function(const char* function_name, const char *class_nam
            if (function_name[10] == 'l') {
             if (function_name[11] == 't') {
              if (function_name[12] == '\0') {
+              ++hp_globals.internal_match_counters.function;
+              const char *class_name = hp_get_class_name(data TSRMLS_CC);
+              if (!class_name) return -1;
               if (!strcmp(class_name, "Magento\\Framework\\View\\Result\\Page\\Interceptor")) return 10;
+              ++hp_globals.internal_match_counters.class_unmatched;
               return -1;
              }
              return -1;
             }
+            return -1;
+           }
+           return -1;
+          }
+          return -1;
+         }
+         return -1;
+        }
+        return -1;
+       }
+       return -1;
+      }
+      return -1;
+     }
+     return -1;
+    }
+    return -1;
+   }
+   if (function_name[2] == 'i') {
+    if (function_name[3] == 'n') {
+     if (function_name[4] == 'd') {
+      if (function_name[5] == 'e') {
+       if (function_name[6] == 'x') {
+        if (function_name[7] == 'A') {
+         if (function_name[8] == 'l') {
+          if (function_name[9] == 'l') {
+           if (function_name[10] == '\0') {
+            ++hp_globals.internal_match_counters.function;
+            const char *class_name = hp_get_class_name(data TSRMLS_CC);
+            if (!class_name) return -1;
+            if (!strcmp(class_name, "Magento\\Indexer\\Model\\Indexer")) return 20;
+            ++hp_globals.internal_match_counters.class_unmatched;
             return -1;
            }
            return -1;
@@ -422,7 +501,11 @@ int hp_match_monitored_function(const char* function_name, const char *class_nam
   if (function_name[1] == 'u') {
    if (function_name[2] == 'n') {
     if (function_name[3] == '\0') {
+     ++hp_globals.internal_match_counters.function;
+     const char *class_name = hp_get_class_name(data TSRMLS_CC);
+     if (!class_name) return -1;
      if (!strcmp(class_name, "Magento\\Framework\\App\\Bootstrap")) return 0;
+     ++hp_globals.internal_match_counters.class_unmatched;
      return -1;
     }
     return -1;
